sweet alret added

+++ src/components/Pages/Dashboard/Admin/ManageService/ManageService.jsx
@@ -1,23 +1,48 @@
 import React, { useEffect, useState } from "react";
+import { useHistory } from "react-router";
+import Swal from "sweetalert2";
 import Spinner from "../../../../Shared/Loader/Spinner";
 import SingleService from "./SingleService";
 
 const ManageService = () => {
   const [services, setServices] = useState([]);
   const [isDelete, setIsDelete] = useState(null);
+  const history = useHistory();
 
-  const deleteOrder = (id) => {
-    fetch(`https://light-wars.herokuapp.com/glasses/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount) {
-          setIsDelete(!isDelete);
-        } else {
-          setIsDelete(false);
-        }
-      });
+  const deleteService = (id) => {
+    Swal.fire({
+      title: "Are you sure?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, Delete!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Successfully added new service",
+          showClass: {
+            popup: "animate__animated animate__fadeInDown",
+          },
+          hideClass: {
+            popup: "animate__animated animate__fadeOutUp",
+          },
+        });
+
+        history.push("/dashboard");
+        fetch(`https://light-wars.herokuapp.com/glasses/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount) {
+              setIsDelete(!isDelete);
+            } else {
+              setIsDelete(false);
+            }
+          });
+      }
+    });
   };
 
   useEffect(() => {
@@ -34,7 +59,7 @@ const ManageService = () => {
         <SingleService
           key={service._id}
           service={service}
-          deleteOrder={deleteOrder}
+          deleteService={deleteService}
         />
       ))}
     </div>
